finished env refactor
