--- src/obsidian_tears_backend/consts.mo
@@ -1,7 +1,0 @@
-module {
-    // public let itemCanisterId : Text = "goei2-daaaa-aaaao-aaiua-cai";
-    public let itemCanisterId : Text = "renrk-eyaaa-aaaaa-aaada-cai";
-
-    // public let characterCanisterId : Text = "dhyds-jaaaa-aaaao-aaiia-cai";
-    public let characterCanisterId : Text = "rkp4c-7iaaa-aaaaa-aaaca-cai";
-};
